+++ src/prime_rl/configs/rl_trainer.py
@@ -4,22 +4,20 @@ from typing import Annotated, Any, Literal, TypeAlias
 from pydantic import BaseModel, Discriminator, Field, Tag, model_validator
 
 from prime_rl.configs.shared import (
-    FileSystemTransportConfig,
-    HeartbeatConfig,
-    LogConfig,
-    MetricsServerConfig,
-    TransportConfigType,
-    WandbConfig,
-)
-from prime_rl.configs.trainer import (
     AdamWConfig,
     BenchConfig,
     CheckpointConfig,
     ConstantSchedulerConfig,
-    ModelConfig,
+    FileSystemTransportConfig,
+    HeartbeatConfig,
+    LogConfig,
+    MetricsServerConfig,
     OptimizerConfigType,
     SchedulerConfigType,
     TokenizerConfig,
+    TrainerModelConfig,
+    TransportConfigType,
+    WandbConfig,
 )
 from prime_rl.utils.pydantic_config import BaseConfig, BaseSettings
 
@@ -148,7 +146,7 @@ class RLTrainerConfig(BaseSettings):
     """Configures the RL trainer"""
 
     # The model configuration
-    model: ModelConfig = ModelConfig()
+    model: TrainerModelConfig = TrainerModelConfig()
 
     # The tokenizer configuration
     tokenizer: TokenizerConfig = TokenizerConfig()
