+++ src/prime_rl/trainer/rl/broadcast/__init__.py
@@ -3,7 +3,7 @@ from pathlib import Path
 import torch
 
 from prime_rl.configs.rl_trainer import WeightBroadcastConfigType
-from prime_rl.configs.trainer import LoRAConfig
+from prime_rl.configs.shared import LoRAConfig
 from prime_rl.trainer.rl.broadcast.base import WeightBroadcast
 from prime_rl.trainer.rl.broadcast.filesystem import FileSystemWeightBroadcast
 from prime_rl.trainer.rl.broadcast.nccl import NCCLWeightBroadcast
